+++ pkg/traceql/engine_metrics.go
@@ -11,6 +11,7 @@ import (
 	"sort"
 	"sync"
 	"time"
+	"unsafe"
 
 	"github.com/grafana/tempo/pkg/tempopb"
 	commonv1proto "github.com/grafana/tempo/pkg/tempopb/common/v1"
@@ -19,7 +20,11 @@ import (
 	"github.com/prometheus/prometheus/model/labels"
 )
 
-const internalLabelBucket = "__bucket"
+const (
+	internalLabelBucket   = "__bucket"
+	maxExemplars          = 100
+	maxExemplarsPerBucket = 2
+)
 
 func DefaultQueryRangeStep(start, end uint64) uint64 {
 	delta := time.Duration(end - start)
@@ -187,9 +192,16 @@ func (ls Labels) String() string {
 	return promLabels.Labels().String()
 }
 
+type Exemplar struct {
+	Labels    Labels
+	Value     float64
+	Timestamp uint64
+}
+
 type TimeSeries struct {
-	Labels Labels
-	Values []float64
+	Labels    Labels
+	Values    []float64
+	Exemplars []Exemplar
 }
 
 // SeriesSet is a set of unique timeseries. They are mapped by the "Prometheus"-style
@@ -241,10 +253,29 @@ func (set SeriesSet) ToProtoDiff(req *tempopb.QueryRangeRequest, rangeForLabels 
 			})
 		}
 
+		exemplars := make([]tempopb.Exemplar, 0, len(s.Exemplars))
+		for _, e := range s.Exemplars {
+			labels := make([]commonv1proto.KeyValue, 0, len(e.Labels))
+			for _, label := range e.Labels {
+				labels = append(labels,
+					commonv1proto.KeyValue{
+						Key:   label.Name,
+						Value: label.Value.AsAnyValue(),
+					},
+				)
+			}
+			exemplars = append(exemplars, tempopb.Exemplar{
+				Labels:      labels,
+				Value:       e.Value,
+				TimestampMs: time.Unix(0, int64(e.Timestamp)).UnixMilli(),
+			})
+		}
+
 		ss := &tempopb.TimeSeries{
 			PromLabels: promLabels,
 			Labels:     labels,
 			Samples:    samples,
+			Exemplars:  exemplars,
 		}
 
 		resp = append(resp, ss)
@@ -263,12 +294,15 @@ type VectorAggregator interface {
 // TODO - for efficiency we probably combine this with VectorAggregator (see todo about CountOverTimeAggregator)
 type RangeAggregator interface {
 	Observe(s Span)
+	ObserveExemplar(float64, uint64, Labels)
 	Samples() []float64
+	Exemplars() []Exemplar
 }
 
 // SpanAggregator sorts spans into series
 type SpanAggregator interface {
 	Observe(Span)
+	ObserveWithExemplar(Span, float64, uint64)
 	Series() SeriesSet
 }
 
@@ -304,10 +338,11 @@ func (c *CountOverTimeAggregator) Sample() float64 {
 
 // StepAggregator sorts spans into time slots using a step interval like 30s or 1m
 type StepAggregator struct {
-	start   uint64
-	end     uint64
-	step    uint64
-	vectors []VectorAggregator
+	start, end, step uint64
+	intervals        int
+	vectors          []VectorAggregator
+	exemplars        []Exemplar
+	exemplarBuckets  *bucketSet
 }
 
 var _ RangeAggregator = (*StepAggregator)(nil)
@@ -319,11 +354,16 @@ func NewStepAggregator(start, end, step uint64, innerAgg func() VectorAggregator
 		vectors[i] = innerAgg()
 	}
 
+	exemplars := make([]Exemplar, 0, maxExemplars)
+
 	return &StepAggregator{
-		start:   start,
-		end:     end,
-		step:    step,
-		vectors: vectors,
+		start:           start,
+		end:             end,
+		step:            step,
+		intervals:       intervals,
+		vectors:         vectors,
+		exemplars:       exemplars,
+		exemplarBuckets: newBucketSet(intervals),
 	}
 }
 
@@ -335,12 +375,32 @@ func (s *StepAggregator) Observe(span Span) {
 	s.vectors[interval].Observe(span)
 }
 
+func (s *StepAggregator) ObserveExemplar(value float64, ts uint64, lbls Labels) {
+	if s.exemplarBuckets.testTotal() {
+		return
+	}
+	interval := IntervalOf(ts, s.start, s.end, s.step)
+	if s.exemplarBuckets.addAndTest(interval) {
+		return
+	}
+
+	s.exemplars = append(s.exemplars, Exemplar{
+		Labels:    lbls,
+		Value:     value,
+		Timestamp: ts,
+	})
+}
+
 func (s *StepAggregator) Samples() []float64 {
 	ss := make([]float64, len(s.vectors))
 	for i, v := range s.vectors {
 		ss[i] = v.Sample()
 	}
 	return ss
+}
+
+func (s *StepAggregator) Exemplars() []Exemplar {
+	return s.exemplars
 }
 
 const maxGroupBys = 5 // TODO - This isn't ideal but see comment below.
@@ -457,9 +517,9 @@ func newGroupingAggregator[F FastStatic, S StaticVals](innerAgg func() RangeAggr
 	}
 }
 
-// Observe the span by looking up its group-by attributes, mapping to the series,
-// and passing to the inner aggregate.  This is a critical hot path.
-func (g *GroupingAggregator[F, S]) Observe(span Span) {
+// getGroupingValues gets the grouping values for the span and stores them in the buffer.
+// Returns false if the span should be dropped.
+func (g *GroupingAggregator[F, S]) getGroupingValues(span Span) bool {
 	// Get grouping values
 	// Reuse same buffer
 	// There is no need to reset, the number of group-by attributes
@@ -475,15 +535,21 @@ func (g *GroupingAggregator[F, S]) Observe(span Span) {
 		v, ok := g.byFunc(span)
 		if !ok {
 			// Totally drop this span
-			return
+			return false
 		}
 		g.buf.vals[len(g.byLookups)] = v
 		g.buf.fast[len(g.byLookups)] = v.MapKey()
 	}
 
+	return true
+}
+
+// getSeries gets the series for the current span.
+// It will reuse the last series if possible.
+func (g *GroupingAggregator[F, S]) getSeries() aggregatorWitValues[S] {
+	// Fast path
 	if g.lastSeries.agg != nil && g.lastBuf.fast == g.buf.fast {
-		g.lastSeries.agg.Observe(span)
-		return
+		return g.lastSeries
 	}
 
 	s, ok := g.series[g.buf.fast]
@@ -495,7 +561,35 @@ func (g *GroupingAggregator[F, S]) Observe(span Span) {
 
 	g.lastBuf = g.buf
 	g.lastSeries = s
+	return s
+}
+
+// Observe the span by looking up its group-by attributes, mapping to the series,
+// and passing to the inner aggregate.  This is a critical hot path.
+func (g *GroupingAggregator[F, S]) Observe(span Span) {
+	if !g.getGroupingValues(span) {
+		return
+	}
+
+	s := g.getSeries()
 	s.agg.Observe(span)
+}
+
+func (g *GroupingAggregator[F, S]) ObserveWithExemplar(span Span, value float64, ts uint64) {
+	if !g.getGroupingValues(span) {
+		return
+	}
+
+	s := g.getSeries()
+	s.agg.Observe(span)
+
+	// Observe exemplar
+	all := span.AllAttributes()
+	lbls := make(Labels, 0, len(all))
+	for k, v := range span.AllAttributes() {
+		lbls = append(lbls, Label{k.String(), v})
+	}
+	s.agg.ObserveExemplar(value, ts, lbls)
 }
 
 // labelsFor gives the final labels for the series. Slower and not on the hot path.
@@ -554,8 +648,9 @@ func (g *GroupingAggregator[F, S]) Series() SeriesSet {
 		labels, promLabels := g.labelsFor(s.vals)
 
 		ss[promLabels] = TimeSeries{
-			Labels: labels,
-			Values: s.agg.Samples(),
+			Labels:    labels,
+			Values:    s.agg.Samples(),
+			Exemplars: s.agg.Exemplars(),
 		}
 	}
 
@@ -574,6 +669,18 @@ func (u *UngroupedAggregator) Observe(span Span) {
 	u.innerAgg.Observe(span)
 }
 
+func (u *UngroupedAggregator) ObserveWithExemplar(span Span, value float64, ts uint64) {
+	u.innerAgg.Observe(span)
+
+	all := span.AllAttributes()
+	lbls := make(Labels, 0, len(all))
+	for k, v := range all {
+		lbls = append(lbls, Label{k.String(), v})
+	}
+
+	u.innerAgg.ObserveExemplar(value, ts, lbls)
+}
+
 // Series output.
 // This is tweaked to match what prometheus does.  For ungrouped metrics we
 // fill in a placeholder metric name with the name of the aggregation.
@@ -582,8 +689,9 @@ func (u *UngroupedAggregator) Series() SeriesSet {
 	l := labels.FromStrings(labels.MetricName, u.name)
 	return SeriesSet{
 		l.String(): {
-			Labels: []Label{{labels.MetricName, NewStaticString(u.name)}},
-			Values: u.innerAgg.Samples(),
+			Labels:    []Label{{labels.MetricName, NewStaticString(u.name)}},
+			Values:    u.innerAgg.Samples(),
+			Exemplars: u.innerAgg.Exemplars(),
 		},
 	}
 }
@@ -621,8 +729,8 @@ func (e *Engine) CompileMetricsQueryRangeNonRaw(req *tempopb.QueryRangeRequest, 
 // CompileMetricsQueryRange returns an evalulator that can be reused across multiple data sources.
 // Dedupe spans parameter is an indicator of whether to expect duplicates in the datasource. For
 // example if the datasource is replication factor=1 or only a single block then we know there
-// aren't duplicates and we can make some optimizations.
-func (e *Engine) CompileMetricsQueryRange(req *tempopb.QueryRangeRequest, dedupeSpans bool, timeOverlapCutoff float64, allowUnsafeQueryHints bool) (*MetricsEvalulator, error) {
+// aren't duplicates, and we can make some optimizations.
+func (e *Engine) CompileMetricsQueryRange(req *tempopb.QueryRangeRequest, dedupeSpans, exemplarsEnabled bool, timeOverlapCutoff float64, allowUnsafeQueryHints bool) (*MetricsEvalulator, error) {
 	if req.Start <= 0 {
 		return nil, fmt.Errorf("start required")
 	}
@@ -657,6 +765,8 @@ func (e *Engine) CompileMetricsQueryRange(req *tempopb.QueryRangeRequest, dedupe
 		metricsPipeline:   metricsPipeline,
 		dedupeSpans:       dedupeSpans,
 		timeOverlapCutoff: timeOverlapCutoff,
+		exemplarsEnabled:  exemplarsEnabled,
+		exemplarBloom:     make(map[string]struct{}, maxExemplars), // TODO: Lazy, use bloom filter, CM sketch or something
 	}
 
 	// TraceID (optional)
@@ -702,6 +812,10 @@ func (e *Engine) CompileMetricsQueryRange(req *tempopb.QueryRangeRequest, dedupe
 	me.start = req.Start
 	me.end = req.End
 
+	if me.exemplarsEnabled {
+		storageReq.SecondPassConditions = append(storageReq.SecondPassConditions, ExemplarMetaConditionsWithout(storageReq.SecondPassConditions)...)
+		storageReq.Exemplars = true
+	}
 	// Setup second pass callback.  It might be optimized away
 	storageReq.SecondPass = func(s *Spanset) ([]*Spanset, error) {
 		// The traceql engine isn't thread-safe.
@@ -786,17 +900,17 @@ func lookup(needles []Attribute, haystack Span) Static {
 }
 
 type MetricsEvalulator struct {
-	start, end        uint64
-	checkTime         bool
-	dedupeSpans       bool
-	deduper           *SpanDeduper2
-	timeOverlapCutoff float64
-	storageReq        *FetchSpansRequest
-	metricsPipeline   metricsFirstStageElement
-	spansTotal        uint64
-	spansDeduped      uint64
-	bytes             uint64
-	mtx               sync.Mutex
+	start, end                      uint64
+	checkTime                       bool
+	dedupeSpans                     bool
+	deduper                         *SpanDeduper2
+	exemplarsEnabled                bool
+	exemplarBloom                   map[string]struct{}
+	timeOverlapCutoff               float64
+	storageReq                      *FetchSpansRequest
+	metricsPipeline                 metricsFirstStageElement
+	spansTotal, spansDeduped, bytes uint64
+	mtx                             sync.Mutex
 }
 
 func timeRangeOverlap(reqStart, reqEnd, dataStart, dataEnd uint64) float64 {
@@ -837,6 +951,8 @@ func (e *MetricsEvalulator) Do(ctx context.Context, f SpansetFetcher, fetcherSta
 		}
 	}
 
+	storageReq.Exemplars = e.exemplarsEnabled
+
 	fetch, err := f.Fetch(ctx, storageReq)
 	if errors.Is(err, util.ErrUnsupported) {
 		return nil
@@ -876,9 +992,13 @@ func (e *MetricsEvalulator) Do(ctx context.Context, f SpansetFetcher, fetcherSta
 			}
 
 			e.spansTotal++
-			e.metricsPipeline.observe(s)
 
+			if e.sampleExemplar(ss.TraceID) {
+				e.metricsPipeline.observe(s, ss.TraceID)
+			}
+			e.metricsPipeline.observe(s, nil)
 		}
+
 		e.mtx.Unlock()
 		ss.Release()
 	}
@@ -899,6 +1019,26 @@ func (e *MetricsEvalulator) Metrics() (uint64, uint64, uint64) {
 
 func (e *MetricsEvalulator) Results() SeriesSet {
 	return e.metricsPipeline.result()
+}
+
+func (e *MetricsEvalulator) sampleExemplar(id []byte) bool {
+	if !e.exemplarsEnabled {
+		return false
+	}
+	if len(e.exemplarBloom) >= maxExemplars {
+		return false
+	}
+	if len(id) == 0 {
+		return false
+	}
+
+	s := unsafe.String(unsafe.SliceData(id), len(id))
+	// Avoid sampling exemplars for the same trace
+	if _, ok := e.exemplarBloom[s]; ok {
+		return false
+	}
+	e.exemplarBloom[s] = struct{}{}
+	return true
 }
 
 // SpanDeduper2 is EXTREMELY LAZY. It attempts to dedupe spans for metrics
@@ -972,17 +1112,20 @@ type SeriesAggregator interface {
 
 type SimpleAdditionAggregator struct {
 	ss               SeriesSet
+	exemplarBuckets  *bucketSet
 	len              int
 	start, end, step uint64
 }
 
 func NewSimpleAdditionCombiner(req *tempopb.QueryRangeRequest) *SimpleAdditionAggregator {
+	l := IntervalCount(req.Start, req.End, req.Step)
 	return &SimpleAdditionAggregator{
-		ss:    make(SeriesSet),
-		len:   IntervalCount(req.Start, req.End, req.Step),
-		start: req.Start,
-		end:   req.End,
-		step:  req.Step,
+		ss:              make(SeriesSet),
+		exemplarBuckets: newBucketSet(l),
+		len:             l,
+		start:           req.Start,
+		end:             req.End,
+		step:            req.Step,
 	}
 }
 
@@ -1000,8 +1143,9 @@ func (b *SimpleAdditionAggregator) Combine(in []*tempopb.TimeSeries) {
 			}
 
 			existing = TimeSeries{
-				Labels: labels,
-				Values: make([]float64, b.len),
+				Labels:    labels,
+				Values:    make([]float64, b.len),
+				Exemplars: make([]Exemplar, 0, len(ts.Exemplars)),
 			}
 			b.ss[ts.PromLabels] = existing
 		}
@@ -1012,6 +1156,35 @@ func (b *SimpleAdditionAggregator) Combine(in []*tempopb.TimeSeries) {
 				existing.Values[j] += sample.Value
 			}
 		}
+
+		for _, exemplar := range ts.Exemplars {
+			if b.exemplarBuckets.testTotal() {
+				break
+			}
+			interval := IntervalOfMs(exemplar.TimestampMs, b.start, b.end, b.step)
+			if b.exemplarBuckets.addAndTest(interval) {
+				continue // Skip this exemplar and continue, next exemplar might fit in a different bucket
+			}
+
+			labels := make(Labels, 0, len(exemplar.Labels))
+			for _, l := range exemplar.Labels {
+				labels = append(labels, Label{
+					Name:  l.Key,
+					Value: StaticFromAnyValue(l.Value),
+				})
+			}
+			value := exemplar.Value
+			if math.IsNaN(value) {
+				value = 0 // TODO: Use the value of the series at the same timestamp
+			}
+			existing.Exemplars = append(existing.Exemplars, Exemplar{
+				Labels:    labels,
+				Value:     value,
+				Timestamp: uint64(time.Duration(exemplar.TimestampMs) * time.Millisecond),
+			})
+		}
+
+		b.ss[ts.PromLabels] = existing
 	}
 }
 
@@ -1052,16 +1225,20 @@ type HistogramAggregator struct {
 	qs               []float64
 	len              int
 	start, end, step uint64
+	exemplars        []Exemplar
+	exemplarBuckets  *bucketSet
 }
 
 func NewHistogramAggregator(req *tempopb.QueryRangeRequest, qs []float64) *HistogramAggregator {
+	l := IntervalCount(req.Start, req.End, req.Step)
 	return &HistogramAggregator{
-		qs:    qs,
-		ss:    make(map[string]histSeries),
-		len:   IntervalCount(req.Start, req.End, req.Step),
-		start: req.Start,
-		end:   req.End,
-		step:  req.Step,
+		qs:              qs,
+		ss:              make(map[string]histSeries),
+		len:             l,
+		start:           req.Start,
+		end:             req.End,
+		step:            req.Step,
+		exemplarBuckets: newBucketSet(l),
 	}
 }
 
@@ -1112,6 +1289,29 @@ func (h *HistogramAggregator) Combine(in []*tempopb.TimeSeries) {
 				existing.hist[j].Record(b, int(sample.Value))
 			}
 		}
+
+		for _, exemplar := range ts.Exemplars {
+			if h.exemplarBuckets.testTotal() {
+				break
+			}
+			interval := IntervalOfMs(exemplar.TimestampMs, h.start, h.end, h.step)
+			if h.exemplarBuckets.addAndTest(interval) {
+				continue // Skip this exemplar and continue, next exemplar might fit in a different bucket
+			}
+
+			labels := make(Labels, 0, len(exemplar.Labels))
+			for _, l := range exemplar.Labels {
+				labels = append(labels, Label{
+					Name:  l.Key,
+					Value: StaticFromAnyValue(l.Value),
+				})
+			}
+			h.exemplars = append(h.exemplars, Exemplar{
+				Labels:    labels,
+				Value:     exemplar.Value,
+				Timestamp: uint64(time.Duration(exemplar.TimestampMs) * time.Millisecond),
+			})
+		}
 	}
 }
 
@@ -1127,8 +1327,9 @@ func (h *HistogramAggregator) Results() SeriesSet {
 			s := labels.String()
 
 			ts := TimeSeries{
-				Labels: labels,
-				Values: make([]float64, len(in.hist)),
+				Labels:    labels,
+				Values:    make([]float64, len(in.hist)),
+				Exemplars: h.exemplars,
 			}
 			for i := range in.hist {